unify datadog tracing now that it is simple
