--- plugins/datadog_tracer/test/samson_datadog_tracer/apm_test.rb
@@ -1,102 +1,0 @@
-# frozen_string_literal: true
-require_relative "../test_helper"
-
-SingleCov.covered!
-
-describe SamsonDatadogTracer::APM do
-  let(:fake_tracer) do
-    Class.new do
-      def self.trace(*)
-        yield
-      end
-    end
-  end
-
-  describe ".trace_method_execution_scope" do
-    with_env DATADOG_TRACER: '1'
-
-    it "skips tracer when disabled" do
-      with_env DATADOG_TRACER: nil do
-        Datadog.expects(:tracer).never
-        SamsonDatadogTracer::APM.trace_execution_scoped("test") { "without tracer" }
-      end
-    end
-
-    it "trigger tracer when enabled" do
-      Rails.stubs(:env).returns("staging")
-      Datadog.expects(:tracer).returns(fake_tracer)
-      SamsonDatadogTracer::APM.trace_execution_scoped("test") { "with tracer" }
-    end
-  end
-
-  describe ".trace_method" do
-    let(:instance) do
-      Class.new do
-        include SamsonDatadogTracer::APM
-
-        def pub_method
-          :pub
-        end
-
-        protected
-
-        def pro_method
-          :pro
-        end
-
-        private
-
-        def pri_method
-          :pri
-        end
-
-        SamsonDatadogTracer::APM.trace_method self, :pub_method
-        SamsonDatadogTracer::APM.trace_method self, :pri_method
-        SamsonDatadogTracer::APM.trace_method self, :pro_method
-      end.new
-    end
-
-    it "wraps methods in a trace call" do
-      Datadog.expects(:tracer).times(3).returns(fake_tracer)
-      instance.send(:pub_method).must_equal(:pub)
-      instance.send(:pro_method).must_equal(:pro)
-      instance.send(:pri_method).must_equal(:pri)
-    end
-
-    it "refuses to add the same wrapper twice since that would lead to infinite loops" do
-      e = assert_raise RuntimeError do
-        SamsonDatadogTracer::APM.trace_method instance.class, :pub_method
-      end
-      e.message.must_include "Tracer already defined for pub_method"
-    end
-
-    [:pub_method, :pro_method, :pri_method].each do |method|
-      it "refuses to add the same wrapper twice for #{method}" do
-        e = assert_raise RuntimeError do
-          SamsonDatadogTracer::APM.trace_method instance.class, method
-        end
-        e.message.must_include "Tracer already defined for #{method}"
-      end
-    end
-
-    it "fails with undefined method" do
-      e = assert_raise NameError do
-        SamsonDatadogTracer::APM.trace_method instance.class, :no_method
-      end
-      e.message.must_include "undefined method `no_method'"
-    end
-
-    it "preserves method visibility" do
-      instance.public_methods.must_include :pub_method
-      instance.public_methods.wont_include :pri_method
-      instance.protected_methods.must_include :pro_method
-      instance.private_methods.must_include :pri_method
-    end
-
-    it "defines alias methods for without" do
-      instance.public_methods.must_include :without_apm_tracer_pub_method
-      instance.protected_methods.must_include :without_apm_tracer_pro_method
-      instance.private_methods.must_include :without_apm_tracer_pri_method
-    end
-  end
-end
